+++ auditor/src/routes/record_handlers.rs
@@ -18,27 +18,22 @@ pub async fn query_records(
 
     let filters: Filters = match serde_qs::from_str(query_string) {
         Ok(filters) => filters,
-        Err(_) => return Err(GetFilterError::InvalidQuery),
+        Err(err) => return Err(GetFilterError::InvalidQuery(err.to_string())),
     };
 
     if query_string.is_empty() {
         // This case explicitly checks if the query is empty. Then it returns all records.
-        let records = advanced_record_filtering(filters, &pool)
-            .await
-            .map_err(|err| GetFilterError::UnexpectedError(err.to_string()))?;
 
-        return Ok(HttpResponse::Ok().json(records));
+        let stream = advanced_record_filtering(filters, pool.as_ref().clone()).await;
+        return Ok(HttpResponse::Ok()
+            .content_type("application/json")
+            .streaming(stream));
     }
 
-    if filters.is_all_none() {
-        return Err(GetFilterError::InvalidQuery);
-    }
-
-    let records = advanced_record_filtering(filters, &pool)
-        .await
-        .map_err(|err| GetFilterError::UnexpectedError(err.to_string()))?;
-
-    Ok(HttpResponse::Ok().json(records))
+    let stream = advanced_record_filtering(filters, pool.as_ref().clone()).await;
+    Ok(HttpResponse::Ok()
+        .content_type("application/json")
+        .streaming(stream))
 }
 
 #[tracing::instrument(name = "Getting one record", skip(record_query, pool))]
@@ -55,7 +50,7 @@ pub async fn query_one_record(
 #[derive(Debug, Error)]
 pub enum GetFilterError {
     #[error("Invalid query parameters")]
-    InvalidQuery,
+    InvalidQuery(String),
 
     #[error("Unexpected error: {0}")]
     UnexpectedError(String),
@@ -64,8 +59,8 @@ pub enum GetFilterError {
 impl ResponseError for GetFilterError {
     fn error_response(&self) -> HttpResponse {
         match self {
-            GetFilterError::InvalidQuery => {
-                HttpResponse::BadRequest().json(json!({ "error": "Invalid query parameters" }))
+            GetFilterError::InvalidQuery(msg) => {
+                HttpResponse::BadRequest().json(json!({ "error": msg }))
             }
             GetFilterError::UnexpectedError(err) => {
                 HttpResponse::InternalServerError().json(json!({ "error": err }))